citiagent_submission/citi/requirements.txt
google-adk==1.3.0
firebase-admin
pandas
numpy
sqlalchemy
pymysql
statsmodels>=0.14
scipy>=1.11
patsy>=0.5
db-dtypes>=1.0.0